# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EarnQA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/by-category:
    get:
      operationId: getAnalyticsByCategory
      tags: [analytics]
      summary: Get answer analytics grouped by category
      parameters:
        - name: questionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Analytics by category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryAnalytics"
  /analytics/by-gender:
    get:
      operationId: getAnalyticsByGender
      tags: [analytics]
      summary: Get answer analytics grouped by gender
      parameters:
        - name: questionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Analytics by gender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenderAnalytics"
  /analytics/by-age:
    get:
      operationId: getAnalyticsByAge
      tags: [analytics]
      summary: Get answer analytics grouped by age group
      parameters:
        - name: questionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Analytics by age
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgeAnalytics"
  /analytics/by-city:
    get:
      operationId: getAnalyticsByCity
      tags: [analytics]
      summary: Get answer analytics grouped by city
      parameters:
        - name: questionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Analytics by city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CityAnalytics"
...
components:
  schemas:
...
    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - category
        - count
        - percentage
    CategoryAnalytics:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
      required:
        - data
    GenderBreakdown:
      type: object
      properties:
        gender:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - gender
        - count
        - percentage
    GenderAnalytics:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GenderBreakdown"
      required:
        - data
    AgeBreakdown:
      type: object
      properties:
        ageGroup:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - ageGroup
        - count
        - percentage
    AgeAnalytics:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AgeBreakdown"
      required:
        - data
    CityBreakdown:
      type: object
      properties:
        city:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - city
        - count
        - percentage
    CityAnalytics:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CityBreakdown"
      required:
        - data
...